delete old edamame stacks in case of aws teardown issues
